+++ user_sync/app.py
@@ -18,29 +18,25 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-import argparse
 import logging
 import os
-import sys
-import click
 import shutil
-from click_default_group import DefaultGroup
+import sys
 from datetime import datetime
-import yaml
-import random
-import string
 
+import click
 import six
+from click_default_group import DefaultGroup
 
+import user_sync.cli
 import user_sync.config
 import user_sync.connector.directory
 import user_sync.connector.umapi
+import user_sync.encryption
 import user_sync.helper
 import user_sync.lockfile
-import user_sync.rules
-import user_sync.cli
 import user_sync.resource
-from user_sync.rsaencryptor import RSAEncryptor
+import user_sync.rules
 from user_sync.certgen import Certgen
 from user_sync.error import AssertionException
 from user_sync.version import __version__ as app_version
@@ -232,8 +228,8 @@ def example_config(**kwargs):
 @click.option('--password', '-p', prompt='Create password', hide_input=True, confirmation_prompt=True)
 def encrypt(password, key_path):
     try:
-        data = RSAEncryptor.encrypt(password, key_path)
-        RSAEncryptor.write_key(data, key_path)
+        data = user_sync.encryption.encrypt(password, key_path)
+        user_sync.encryption.write_key(data, key_path)
         click.echo('Encryption was successful.\n{0}'.format(os.path.abspath(key_path)))
     except AssertionException as e:
         click.echo(str(e))
@@ -244,38 +240,14 @@ def encrypt(password, key_path):
 @click.option('--password', '-p', prompt='Enter password', hide_input=True)
 def decrypt(password, key_path):
     try:
-        data = RSAEncryptor.decrypt(password, key_path)
-        RSAEncryptor.write_key(data, key_path)
+        data = user_sync.encryption.decrypt(password, key_path)
+        user_sync.encryption.write_key(data, key_path)
         click.echo('Decryption was successful.\n{0}'.format(os.path.abspath(key_path)))
     except AssertionException as e:
         click.echo(str(e))
 
 
-@main.command(help='Create a new certificate_pub.crt file and private.key file. '
-                   'User Sync Tool can use these certificates to communicate with '
-                   'the admin console. Please visit https://console.adobe.io to '
-                   'complete the integration process.')
-@click.option('--overwrite', '-y', help='Overwrite files without being asked to confirm', is_flag=True)
-@click.option('--randomize', '-r', help='Randomize the values rather than entering credentials', is_flag=True)
-@click.option('--private-key-file', '-p', help='Set a custom path to a private.key file', default='private.key')
-@click.option('--cert-pub-file', '-c', help='Set a custom path to a certificate_pub.crt file', default='certificate_pub.crt')
-def certgen(randomize, private_key_file, cert_pub_file, overwrite):
-    private_key_file = os.path.abspath(private_key_file)
-    cert_pub_file = os.path.abspath(cert_pub_file)
-    if not overwrite and (os.path.exists(private_key_file) or os.path.exists(cert_pub_file)):
-        if not click.confirm('Would you like to overwrite the original files?'):
-            return
-    try:
-        if not randomize:
-            click.echo("Enter information as required to generate the X509 certificate/key pair for your organization. "
-                       "This information is used only for authentication with UMAPI and does not need to reflect "
-                       "an SSL or other official identity.")
-        subject_fields = Certgen.get_subject_fields(randomize)
-        Certgen.generate(private_key_file, cert_pub_file, subject_fields)
-        click.echo("Files were created at:\n{0}\n{1}".format(private_key_file, cert_pub_file))
-    except AssertionException as e:
-        click.echo(str(e))
-        click.echo('Files have not been created/overwritten.')
+
 
 
 @main.command()
